refactor(add-transaction): split the per-asset buy forms into focused components (#467)
AddTransactionSheet's render was a ~1180-line monolith. Extract the three BUY
asset field sections — fund, bank, gold — into `addTransactionForms.tsx` as
`BuyFundFields` / `BuyBankFields` / `BuyGoldFields`. The sheet keeps the
orchestration (state, effects, dir/assetType gating, submit) and passes each
asset's state down; the form components own only their inputs and pull `t` via
their own `useTranslations`. Business math stays in addTransactionModel.ts;
`GOLD_PROVIDERS` moves with the gold form. No behavior or markup change.

This is the first split of AddTransactionSheet (#467's "large files split along
clear boundaries"); the sell workflow is a natural follow-up.

Parity is guarded by the existing AddTransactionSheet component tests (which
exercise fund NAV prefill + unit derivation, the bank fields, and gold buy) plus
the add-transaction and accumulating-deposit E2E — both green — so the extracted
JSX renders and wires identically.

Verified: AddTransactionSheet 24 + full unit suite 1250 passed; typecheck 0
errors; lint 0 errors; E2E add-transaction (14) + accumulating-deposit (3) passed.
AddTransactionSheet drops 1189 → 959 lines.
